:recycle: (refactor) rename chatbot services
